View part from Use case is finished
update account model
update database access class
Create account server access in client
Create account service in client
Create account viewing view model
Create account viewing fxml page
Create account viewing controller

+++ MoneyTrackClient/src/mvvm/model/Validator.java
@@ -31,14 +31,14 @@ public class Validator {
         return isValidEmail(email)&&email.contains(ADMIN_DOMAIN);
     }
 
-    public static boolean isValidIncome(int amount){
+    public static boolean isValidIncome(float amount){
         return amount>=0&&amount<=MAX_INCOME;
     };
-    public static boolean isValidExpense(int amount){
+    public static boolean isValidExpense(float amount){
         return amount>=0&&amount<=MAX_EXPENSE;
 
     };
-    public static boolean isValidBalance(int amount){
+    public static boolean isValidBalance(float amount){
         return amount>=0&&amount<=MAX_START_BALANCE;
     };
     public static boolean isValidPriceInEur(double amount){
